# coral: add static linting