--- a/coral/package.json
+++ b/coral/package.json
@@ -12,7 +12,9 @@
   "scripts": {
     "dev": "vite",
     "build": "tsc && vite build",
-    "preview": "vite preview"
+    "preview": "vite preview",
+    "lint": "prettier --check src",
+    "reformat": "prettier --write src"
   },
   "dependencies": {
     "react": "^18.2.0",
@@ -22,6 +24,7 @@
     "@types/react": "^18.0.17",
     "@types/react-dom": "^18.0.6",
     "@vitejs/plugin-react": "^2.1.0",
+    "prettier": "^2.7.1",
     "typescript": "^4.6.4",
     "vite": "^3.1.0"
   }