--- a/coral/src/App.tsx
+++ b/coral/src/App.tsx
@@ -1,10 +1,9 @@
 function App() {
-
   return (
     <div>
       <h1>Hello Klaw 👋</h1>
     </div>
-  )
+  );
 }
 
-export default App
+export default App;