Introduce the following commands:
- `pnpm run lint`
  - Runs both [prettier](https://prettier.io/) check and [ESLint](https://eslint.org/)
- `pnpm run reformat`
  - Runs [prettier](https://prettier.io/) with `--write` flag
  
The current configuration for Prettier and ESLint are minimal defaults.

Resolves: #110