@SmuliS 2022-10-19
I would need to do some more research on the topic. I have a vague memory that when you bundle the tools together it kills the overall performance. Below are results from some benchmarking:

Without `eslint-plugin-prettier`:
```
smulis@mbp coral % pnpm run lint

> coral@0.0.1 lint /Users/smulis/git/klaw/coral
> TIMING=1 eslint .

Rule                                      | Time (ms) | Relative
:-----------------------------------------|----------:|--------:
react/display-name                        |     5.856 |    34.5%
@typescript-eslint/no-unused-vars         |     4.271 |    25.2%
react/require-render-return               |     0.551 |     3.2%
react/no-direct-mutation-state            |     0.474 |     2.8%
react/prop-types                          |     0.409 |     2.4%
@typescript-eslint/no-empty-function      |     0.296 |     1.7%
no-misleading-character-class             |     0.277 |     1.6%
@typescript-eslint/triple-slash-reference |     0.272 |     1.6%
react/no-deprecated                       |     0.256 |     1.5%
no-global-assign                          |     0.247 |     1.5%
```

With `eslint-plugin-prettier`:
```
smulis@mbp coral % pnpm run lint

> coral@0.0.1 lint /Users/smulis/git/klaw/coral
> TIMING=1 eslint .

Rule                                      | Time (ms) | Relative
:-----------------------------------------|----------:|--------:
prettier/prettier                         |   228.112 |    94.2%
react/display-name                        |     4.266 |     1.8%
@typescript-eslint/no-unused-vars         |     2.923 |     1.2%
react/no-direct-mutation-state            |     0.558 |     0.2%
react/require-render-return               |     0.503 |     0.2%
react/prop-types                          |     0.422 |     0.2%
@typescript-eslint/no-empty-function      |     0.337 |     0.1%
@typescript-eslint/triple-slash-reference |     0.302 |     0.1%
@typescript-eslint/ban-types              |     0.277 |     0.1%
no-misleading-character-class             |     0.271 |     0.1%
```

There is also this closed issue in the repository. Not sure if applies anymore:
https://github.com/prettier/eslint-plugin-prettier/issues/304