@programmiri 2022-10-19
😱  Ok that's some serious difference, I was not aware of that. 

Is the overall run of `lint` in comparison also with such a big difference? As far as I understood `TIMING=1 eslint` benchmarks the eslint rules, right? I did try `time pnpm lint` for both settings and the difference is much smaller (1.20s with prettier, 1.09s with eslint-plugin-prettier). 

In general I don't have strong opinions on either way, I just remember that my IDE got very confused with the combination (running eslint and prettier on save), but that's probably a setup problem 🤔