@SmuliS 2022-10-20
Updated the PR to include: [eslint-config-prettier](https://github.com/prettier/eslint-config-prettier). This is ready to be reviewed again.